==> 0_install_packages.R <==
# install packages needed for the analysis

install.packages("Seurat")
install.packages("BiocManager")
BiocManager::install("edgeR")     # Install edgeR using BiocManager
install.packages("gplots")  # Install gplots if not already installed
BiocManager::install("SeuratWrappers")
install.packages("dplyr")
install.packages("cowplot")
install.packages("gridExtra")
install.packages("cluster")
install.packages("mgcv")
install.packages("Matrix")
install.packages("clue")
install.packages("igraph")
install.packages("tidyr")
install.packages("openxlsx")
install.packages("ggrepel")
install.packages("patchwork")
